requirements.txt
asgiref==3.6.0
asyncpg==0.27.0
databases==0.7.0
Django==4.1.5
django-ash==0.0.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
greenlet==2.0.1
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
SQLAlchemy==1.4.46
sqlparse==0.4.3
